+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/KafkaMirrorMaker2ClusterTest.java
@@ -1625,7 +1625,7 @@ public class KafkaMirrorMaker2ClusterTest {
                             .withConnectTimeoutSeconds(15)
                             .withReadTimeoutSeconds(15)
                             .withHttpRetries(2)
-                            .withHttpRetryPauseMillis(500)
+                            .withHttpRetryPauseMs(500)
                             .withEnableMetrics(true)
                             .withNewRefreshToken()
                                 .withSecretName("my-token-secret")